stdlib: Updating dma_requestor for chi.
This change updates the configuration of dma_requestor in chi
cachehierachy. It sets two flags of `dealloc_backinv_unique` and
`dealloc_backinv_shared` to False. Since the dma_requestor does
not have an upstream cache, it does not really matter whether
they are set to true or false. However, setting them to false
ensures the unearthing of any flaws in the underlying SLICC code.

Change-Id: Ifb185eecbb6ff10657d64715e3fde020e71f3d9d